Added general_utilities.py file.
In this commit The general_utilities.py file contains the function to
calculate the dimensions of output of each layer of a CNN.
There are no parameters passed to this fucntion during
function call, instead parameters are prompted through
user input. More functons or classes maybe added to this file in
the future.